+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 1.5.5 - Unreleased
 
+### Added
+* `set_overlaps` now has an `n_samples` keyword argument.
+* Add `integer` pattern to regex builder
+
 ### Change
 * freshpyenv now uses uv